refactor: restructure schema validations
