+++ src/stack/index.ts
@@ -1,3 +1,2 @@
 export { deployStack } from './deploy';
 export * from './parse';
-export * from './iacSchema';
