+++ src/types/domains/vars.ts
