+++ src/validations/index.ts
@@ -1,0 +1,1 @@
+export { validateYaml } from './iacSchema';
