+++ tests/validations/validate.test.ts
@@ -1,8 +1,8 @@
 import { readFileSync } from 'node:fs';
 import * as path from 'node:path';
 import { parse } from 'yaml';
-import { validateYaml } from '../../src/stack';
 import { ServerlessIacRaw } from '../../src/types';
+import { validateYaml } from '../../src/validations';
 
 const jsonIac = parse(
   readFileSync(path.resolve(__dirname, '../fixtures/serverless-insight.yml'), 'utf8'),
